OpenvSwitch workaround for wrong number of bits sent in some table_stats bitmaps.
OVS was sending bitmaps for the match,wildcards,write_set and apply_set
containing 1.3 bits. This was reported to OVS and a patch is in the works.
This fix however make sure to clean out any preceeding bits of the max
allowed (36 bits) number of bits according to the 1.2 spec.